[fix] 정리된 파일에 맞게 채팅 수정
