+++ src/pages/chat.html
@@ -6,7 +6,7 @@
     <title>채팅 | A13C</title>
     <link rel="icon" type="image/svg+xml" href="/logo.svg" />
     <script type="module" src="/src/main.ts"></script>
-    <script type="module" src="/src/pages/ingame/chat.ts"></script>
+    <script type="module" src="/src/script/ingame/chat.ts"></script>
   </head>
   <body class="bg-cyan-800">
     <h1 class="text-2xl">채팅</h1>
