+++ src/pages/ingame.html
@@ -9,6 +9,9 @@
     <link rel="stylesheet" href="/font/pretendardvariable.css" />
     <script type="module" src="/src/main.ts"></script>
     <script type="module" src="/src/script/ingame/index.ts"></script>
+    <script type="module" src="/src/script/A13C-chat.ts"></script>
+    <script type="module" src="/src/script/ingame/chat.ts"></script>
+
 
     <!-- <script type="module" src="/src/pages/ingame/ingame-ui.ts"></script> -->
     <!-- <script src="https://fesp-api.koyeb.app/socket.io/socket.io.js"></script> -->
