--- src/pages/ingame/ingame.ts
@@ -1,1 +1,0 @@
-import "./ingame-server";
