+++ src/pages/lobby.html
@@ -8,6 +8,7 @@
     <link rel="stylesheet" href="/font/pretendardvariable.css" />
     <link rel="icon" type="image/svg+xml" href="/logo.svg" />
     <script type="module" src="/src/script/lobby/index.ts"></script>
+    <script type="module" src="/src/script/A13C-chat.ts"></script>
   </head>
   <body class="font-pretendard">
     <!-- 스크린 리더 사용자를 위한 페이지 소개 시작 -->
